moving tUI shared utillity functions to textUI.h
