+++ JoyReceiver_tUI/JoyReceiver_tUI.h
@@ -26,46 +26,11 @@ THE SOFTWARE.
 #include "./../JoyReceiver++/JoyReceiver++.h"
 #include <codecvt>
 #include <thread>
-#include <mutex>
-std::mutex mtx;  // Mutex for synchronizing access to shared variables
 
 #undef min
 #undef max
 
 #define APP_VERSION_NUM     L"2.0.0.0"
-
-std::string g_outputText;
-
-// Function safely return environment variables
-std::string g_getenv(const char* variableName) {
-    std::string value;
-    size_t requiredSize = 0;
-    getenv_s(&requiredSize, nullptr, 0, variableName);
-    if (requiredSize > 0) {
-        value.resize(requiredSize);
-        getenv_s(&requiredSize, &value[0], requiredSize, variableName);
-        value.pop_back(); // Remove the null-terminating character
-    }
-    return value;
-}
-// Function to detect if keyboard key is pressed
-bool getKeyState(int KEYCODE) {
-    return GetAsyncKeyState(KEYCODE) & 0x8000;
-}
-// Function determines if app is the active window
-bool IsAppActiveWindow()
-{
-    HWND consoleWindow = GetConsoleWindow();
-    HWND foregroundWindow = GetForegroundWindow();
-
-    return (consoleWindow == foregroundWindow);
-}
-// Function to convert strings to wide strings
-std::wstring g_toWide(std::string& str) {
-    std::wstring wideString(str.begin(), str.end());
-    return wideString;
-}
-
 
 #define AXIS_INPUT_DEADZONE   3000    // stick input values must be greater than this to register
 // stripped down SDLButtonMapping from JoySender++
@@ -321,17 +286,10 @@ private:
     }
 };
 
-
 ///////
-/// tUI.h 
+/// textUI 
 //////
 #include "./../JoySender_tUI/tUI/textUI.h"
-
-const int consoleWidth = 72;
-const int consoleHeight = 20;
-
-HANDLE g_hConsoleInput;
-textUI g_screen;
 
 void setErrorMsg(const wchar_t* text, size_t length);
 void updateFPS(const wchar_t* text, size_t length);
@@ -340,36 +298,15 @@ void newControllerColorsCallback(mouseButton& button);
 void checkForQuit();
 bool checkKey(int key, bool pressed);
 int screenLoop(textUI& screen);
-
 void loopCount(int& counter, int maxCount);
 void countUpDown(int& counter, int maxCount);
-
-#define IS_PRESSED      1
-#define IS_RELEASED     0
-std::unordered_map<int, bool> g_keyStateMap;
-// function checks for KEYCODE state & logs to dict. returns true when state change mirrors the bool pressed
-bool checkKey(int key, bool pressed) {
-    bool state = GetAsyncKeyState(key) & 0x8000;
-    if (state != g_keyStateMap[key]) {
-        g_keyStateMap[key] = state;
-        if (state) {
-            // key is pressed
-            if (pressed)
-                return 1;
-        }
-        else {
-            // key is released
-            if (!pressed)
-                return 1;
-        }
-    }
-    return 0;
-}
 
 #pragma warning(disable : 4996)
 // Converts strings to wide strings and back again
 std::wstring_convert<std::codecvt_utf8<wchar_t>> g_converter;
 
+HANDLE g_hConsoleInput;
+textUI g_screen;
 int g_mode = 1;
 int g_currentColorScheme;
 ColorScheme g_simpleScheme;
@@ -381,7 +318,6 @@ wchar_t* fpsPointer = clientPointer + 25;   // max length 10
 wchar_t* msgPointer1 = fpsPointer + 10;   // max length 100
 wchar_t* msgPointer2 = msgPointer1 + 100;   // max length 100
 wchar_t* msgPointer3 = msgPointer2 + 100;   // max length 165
-
 
 textBox errorOut(4, 22, 45, 0, ALIGN_CENTER, errorPointer, BRIGHT_RED);
 textBox clientMsg(1, 8, 20, 0, ALIGN_LEFT, clientPointer, BRIGHT_GREEN);
@@ -402,205 +338,6 @@ void updateFPS(const wchar_t* text, size_t length) {
     wcsncpy_s(fpsPointer, length, text, _TRUNCATE);
     fpsMsg.SetText(fpsPointer);
     fpsMsg.Draw();
-}
-
-
-// ********************************
-// FullColourScheme
-
-// Package up all the colors needed for maximum color expression
-struct FullColorScheme {
-    const wchar_t* name;
-    tUIColorPkg menuColors;
-    tUIColorPkg controllerColors;
-    WORD menuBg;
-    WORD controllerBg;
-};
-
-#define NUM_COLOR_SCHEMES   5   // number of defined color schemes + a place for a random scheme
-#define RANDOMSCHEME        0   // index to store a randomly generated color scheme
-
-FullColorScheme fullColorSchemes[NUM_COLOR_SCHEMES + 1] = {
-    { L"Random",
-        {
-        // Menu Colors
-        WHITE | (MAGENTA AS_BG),			// Header
-        WHITE | (RED AS_BG),		        // Error
-        WHITE | (CYAN AS_BG),				// message 1
-        WHITE | (BRIGHT_RED AS_BG)			// message 2
-    },
-    {
-        // Controller Colors
-        BLACK | (BRIGHT_MAGENTA AS_BG),	    // outline | faceColor 
-        WHITE | (BRIGHT_MAGENTA AS_BG),		// buttonColor
-        BLUE | (BRIGHT_MAGENTA AS_BG),	    // highlightColor
-        WHITE | (CYAN AS_BG)	            // selectColor
-    },
-    // Menu background
-    WHITE | (BRIGHT_RED AS_BG),
-    // Controller background
-    BLACK | (MAGENTA AS_BG),
-},
-
-
-{ L"Plum",
-    {
-        // Menu Colors
-        WHITE | (MAGENTA AS_BG),			// Header
-        WHITE | (RED AS_BG),		        // Error
-        WHITE | (CYAN AS_BG),				// message 1
-        WHITE | (BRIGHT_RED AS_BG)			// message 2
-    },
-        {
-            // Controller Colors
-            BLACK | (BRIGHT_MAGENTA AS_BG),	    // outline | faceColor 
-            WHITE | (BRIGHT_MAGENTA AS_BG),		// buttonColor
-            BLUE | (BRIGHT_MAGENTA AS_BG),	    // highlightColor
-            WHITE | (CYAN AS_BG)	            // selectColor
-        },
-    // Menu background
-    WHITE | (BRIGHT_RED AS_BG),
-    // Controller background
-    BLACK | (MAGENTA AS_BG),
-},
-
-
- { L"Watermelon",
-    {
-        // Menu Colors
-        BLACK | (GREEN AS_BG),			// Header
-        WHITE | (RED AS_BG),		    // Error
-        BLACK | (BRIGHT_CYAN AS_BG),	// message 1
-        BLUE | (BRIGHT_GREEN AS_BG)		// message 2
-    },
-        {
-            // Controller Colors
-            BRIGHT_GREEN | (BRIGHT_RED AS_BG),	// outline | faceColor 
-            BLACK | (BRIGHT_RED AS_BG),		    // buttonColor
-            BRIGHT_GREEN | (BRIGHT_RED AS_BG),  // highlightColor
-            BLACK | (BRIGHT_GREEN AS_BG)        // selectColor
-        },
-    // Menu background
-    BLACK | (CYAN AS_BG),
-    // Controller background
-    BRIGHT_GREEN | (BRIGHT_BLUE AS_BG),
-},
-
-
- { L"Blueberry",
-    {
-        // Menu Colors
-        WHITE | (BRIGHT_MAGENTA AS_BG),		// Header
-        WHITE | (RED AS_BG),		        // Error
-        BLUE | (BRIGHT_CYAN AS_BG),			// message 1
-        BRIGHT_CYAN | (MAGENTA AS_BG)		// message 2
-    },
-        {
-            // Controller Colors
-            CYAN | (CYAN AS_BG),	        // outline | faceColor 
-            WHITE | (CYAN AS_BG),		    // buttonColor
-            BLUE | (CYAN AS_BG),            // highlightColor
-            BRIGHT_CYAN | (BRIGHT_MAGENTA AS_BG)  // selectColor
-        },
-    // Menu background
-    CYAN | (MAGENTA AS_BG),
-    // Controller background
-    CYAN | (BLUE AS_BG),
-},
-
-
-{ L"Citrus",
-   {
-        // Menu Colors
-        BRIGHT_BLUE | (BRIGHT_GREEN AS_BG),	// Header
-        WHITE | (RED AS_BG),		        // Error
-        BLACK | (BRIGHT_CYAN AS_BG),	// message 1
-        WHITE | (CYAN AS_BG)			    // message 2
-    },
-    {
-        // Controller Colors
-        BRIGHT_GREEN | (BRIGHT_GREEN AS_BG),	// outline | faceColor 
-        BLUE | (BRIGHT_GREEN AS_BG),		    // buttonColor
-        WHITE | (BRIGHT_GREEN AS_BG),           // highlightColor
-        WHITE | (YELLOW AS_BG)      // selectColor
-    },
-    // Menu background
-    BRIGHT_BLUE | (BRIGHT_GREEN AS_BG),
-    // Controller background
-    BRIGHT_GREEN | (BRIGHT_RED AS_BG),
-},
-
-
-{ L"Emily's Bougie Baby",
-    {
-        // Menu Colors
-        BLACK | (YELLOW AS_BG),			// Header
-        WHITE | (RED AS_BG),		    // Error
-        WHITE | (BLUE AS_BG),			// message 1
-        BLUE | (BLACK AS_BG)			// message 2
-    },
-        {
-            // Controller Colors
-            YELLOW | (BLACK AS_BG),	        // outline | faceColor 
-            YELLOW | (BLACK AS_BG),		    // buttonColor
-            BRIGHT_YELLOW | (BLACK AS_BG),  // highlightColor
-            WHITE | (YELLOW AS_BG)  // selectColor
-        },
-    // Menu background
-    RED | (BLACK AS_BG),
-    // Controller background
-    YELLOW | (BLACK AS_BG),
-}
-
-};
-
-// takes a color package representing controller button colors and converts to screen button colors for mouse hovering
-tUIColorPkg controllerButtonsToScreenButtons(tUIColorPkg& inButs) {
-    return tUIColorPkg(
-        inButs.col2,    // default
-        inButs.col3,    // highlight
-        inButs.col4,    // select
-        inButs.col3     // active // used for ip input
-    );
-}
-
-// converts a simple color scheme to a full color scheme *random color scheme generation can be converted to a full scheme
-FullColorScheme fullSchemeFromSimpleScheme(ColorScheme& simp, WORD bg) {
-    FullColorScheme ret = {
-        L"Random",
-        {
-            // Menu Colors
-            BLACK | (GREY AS_BG),			// Header
-            BLACK | (RED AS_BG),		        // Error
-            BLACK | (BRIGHT_GREY AS_BG),		// message 1
-            BLACK | (WHITE AS_BG)			// message 2
-        },
-        {
-            // Controller Colors
-            static_cast<WORD>(simp.outlineColor | simp.faceColor),	    // outline | faceColor 
-            static_cast<WORD>(simp.buttonColor | simp.faceColor),		// buttonColor
-            static_cast<WORD>(simp.highlightColor | simp.faceColor),	    // highlightColor
-            simp.selectColor                         // selectColor 
-        },
-        // Menu background
-        BLACK | (WHITE AS_BG),
-        // Controller background
-        static_cast<WORD>(simp.outlineColor | bg)
-    };
-
-    return ret;
-}
-
-// converts a a full color scheme to a simple scheme *needed for draw and redraw controller face functions
-ColorScheme simpleSchemeFromFullScheme(FullColorScheme& full) {
-    return ColorScheme{
-        static_cast<WORD>(full.controllerColors.col1 FG_ONLY),
-        static_cast<WORD>(full.controllerColors.col1 BG_ONLY),
-        static_cast<WORD>(full.controllerColors.col2 FG_ONLY),
-        static_cast<WORD>(full.controllerColors.col3 FG_ONLY),
-        static_cast<WORD>(full.controllerColors.col4),
-        L" "
-    };
 }
 
  
@@ -818,52 +555,6 @@ void buttonStatesFromXboxReport(XUSB_REPORT& xboxReport) {
     }
 }
 
-// checks if the Q button has been pressed and sets APP_KILLED to true
-void checkForQuit() {
-    if (IsAppActiveWindow() && checkKey('Q', IS_PRESSED) && getKeyState(VK_SHIFT)) // Q for Quit
-        APP_KILLED = true;
-}
-
-// used in a loop, looks for mouse input and compares to buttons on screen
-int screenLoop(textUI& screen) {
-    int retVal = 0;
-    DWORD eventsAvailable;
-    GetNumberOfConsoleInputEvents(g_hConsoleInput, &eventsAvailable);
-
-    if (eventsAvailable > 0) {
-        INPUT_RECORD inputRecord;
-        DWORD eventsRead;
-
-        if (!ReadConsoleInput(g_hConsoleInput, &inputRecord, 1, &eventsRead)) {
-            return -1;
-        }
-
-        // Check for mouse input events
-        if (inputRecord.EventType == MOUSE_EVENT) {
-            MOUSE_EVENT_RECORD& mouseEvent = inputRecord.Event.MouseEvent;
-            COORD mousePos = mouseEvent.dwMousePosition;
-
-            if (mouseEvent.dwEventFlags == MOUSE_MOVED) {
-                // Check if mouse is hovering over the buttons
-                screen.CheckMouseHover(mousePos);
-            }
-
-            if (mouseEvent.dwEventFlags == 0) {
-                // Mouse button event
-                if (mouseEvent.dwButtonState == FROM_LEFT_1ST_BUTTON_PRESSED) {
-                    screen.CheckMouseClick(MOUSE_DOWN);
-                    retVal = MOUSE_DOWN;
-                }
-                else {
-                    screen.CheckMouseClick(MOUSE_UP);
-                    retVal = MOUSE_UP;
-                }
-            }
-        }
-    }
-    return retVal;
-}
-
 // to allow for screen loop and animation, wait for connection to client in a separate thread
 void threadedAwaitConnection(TCPConnection& server, int& retVal, char* clientIP) {
     
@@ -919,89 +610,4 @@ void threadedFrameAdvance(int& run, int delay, int& counter, int maxCount) {
         Sleep(delay);
         loopCount(counter, maxCount);
     }
-}
-
-
-///  
-/// ANIMATIONS
-///  
-int g_frameNum = 0;  //
-const int CX_ANI_FRAME_COUNT = 12;
-
-const wchar_t* ConnectAnimationLeft[CX_ANI_FRAME_COUNT + 1] = {
-    {L"     "},
-    {L"   . "},
-    {L"  .. "},
-    {L" ... "},
-    {L" ... "},
-    {L"  .. "},
-    {L"   . "},
-    {L"     "},
-    {L" * * "},
-    {L"  *  "},
-    {L" * * "},
-    {L"  *  "},
-    {L" *** "},
-};
-
-const wchar_t* ConnectAnimationRight[CX_ANI_FRAME_COUNT + 1] = {
-    {L"     "},
-    {L" .   "},
-    {L" ..  "},
-    {L" ... "},
-    {L" ... "},
-    {L" ..  "},
-    {L" .   "},
-    {L"     "},
-    {L" * * "},
-    {L"  *  "},
-    {L" * * "},
-    {L"  *  "},
-    {L" *** "},
-};
-
-const int FOOTER_ANI_FRAME_COUNT = 9;
-const wchar_t* FooterAnimation[CX_ANI_FRAME_COUNT + 1] = {
-{L" .--.     :::::.\\:::'      `--"},
-{L".--.      ::::.\\::::      `--'"},
-{L"--.      .:::.\\:::::     `--' "},
-{L"-.      .-::.\\::::::    `--'  "},
-{L".      .--:.\\:::::::   `--'   "},
-{L"      .--..\\::::::::  `--'    "},
-{L"     .--. \\::::::::. `--'     "},
-{L"    .--.  ::::::::.\\`--'      "},
-{L"   .--.   :::::::.\\:--'      `"},
-{L"  .--.    ::::::.\\::-'      `-"},
-};
-
-// function will inc/dec &counter up to maxCount and down to 0
-// for animating back and forth through frames
-void countUpDown(int& counter, int maxCount) { // * a counter value outside of (0 - maxCount) will cause issues
-    static int dir = 1; // 1: up, -1: down
-
-    // this non intuitive block makes sure we always count up from 0
-    if (!counter) {
-        dir = -1;
-    }
-
-    // if 0 or maxCount, direction of counting will be reversed
-    if (!(counter < maxCount && counter > 0)) {
-        dir *= -1;
-    }
-
-    counter += dir; // count by one in direction +1 or -1
-}
-
-// function will loop from 0 to maxCount
-// for animation cyclically through frames
-void loopCount(int& counter, int maxCount) {
-    counter = (counter + 1) % (maxCount + 1);
-}
-
-// function will loop from maxCount to 0
-// for animation cyclically through frames
-void revLoopCount(int& counter, int maxCount) {
-    --counter;
-    if (counter < 0)
-        counter = maxCount-1;
 }
